+++ mlrun/common/schemas/common.py
@@ -38,46 +38,3 @@ class ImageBuilder(pydantic.BaseModel):
     build_pod: typing.Optional[str] = None
     requirements: typing.Optional[list] = None
     source_code_target_dir: typing.Optional[str] = None
-
-
-class ObjectFormat:
-    full = "full"
-
-    @staticmethod
-    def format_method(_format: str) -> typing.Optional[typing.Callable]:
-        return {
-            ObjectFormat.full: None,
-        }[_format]
-
-    @classmethod
-    def format_obj(cls, obj: dict, _format: str) -> dict:
-        _format = _format or cls.full
-        format_method = cls.format_method(_format)
-        if not format_method:
-            return obj
-
-        return format_method(obj)
-
-    @staticmethod
-    def filter_obj_method(_filter: list[list[str]]) -> typing.Callable:
-        def _filter_method(obj: dict) -> dict:
-            formatted_obj = {}
-            for key_list in _filter:
-                obj_recursive_iterator = obj
-                formatted_obj_recursive_iterator = formatted_obj
-                for idx, key in enumerate(key_list):
-                    if key not in obj_recursive_iterator:
-                        break
-                    value = (
-                        {} if idx < len(key_list) - 1 else obj_recursive_iterator[key]
-                    )
-                    formatted_obj_recursive_iterator.setdefault(key, value)
-
-                    obj_recursive_iterator = obj_recursive_iterator[key]
-                    formatted_obj_recursive_iterator = formatted_obj_recursive_iterator[
-                        key
-                    ]
-
-            return formatted_obj
-
-        return _filter_method
