+++ server/api/api/endpoints/functions.py
@@ -30,6 +30,7 @@ from fastapi.concurrency import run_in_threadpool
 from kubernetes.client.rest import ApiException
 from sqlalchemy.orm import Session
 
+import mlrun.common.formatters
 import mlrun.common.model_monitoring
 import mlrun.common.model_monitoring.helpers
 import mlrun.common.schemas
@@ -113,7 +114,7 @@ async def get_function(
     name: str,
     tag: str = "",
     hash_key="",
-    _format: str = Query(mlrun.common.schemas.FunctionFormat.full, alias="format"),
+    _format: str = Query(mlrun.common.formatters.FunctionFormat.full, alias="format"),
     auth_info: mlrun.common.schemas.AuthInfo = Depends(deps.authenticate_request),
     db_session: Session = Depends(deps.get_db_session),
 ):
@@ -208,7 +209,7 @@ async def list_functions(
     page: int = Query(None, gt=0),
     page_size: int = Query(None, alias="page-size", gt=0),
     page_token: str = Query(None, alias="page-token"),
-    _format: str = Query(mlrun.common.schemas.FunctionFormat.full, alias="format"),
+    _format: str = Query(mlrun.common.formatters.FunctionFormat.full, alias="format"),
     auth_info: mlrun.common.schemas.AuthInfo = Depends(deps.authenticate_request),
     db_session: Session = Depends(deps.get_db_session),
 ):
